+++ CHANGES.md
@@ -5,6 +5,7 @@
 - Added a box-shadow effect to all drop-down context menus.
 - Removed deprecated CSS values.
 - When the browser window loses focus, all drop-down context menus, tooltips, and date pickers, are closed.
+- When moving a dialog, if the mouse leaves the document area, the move will be canceled and the dialogs location will revert back.
 - 
 
 <br>